% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx_helpers.R
\name{as_pretty_tibble}
\alias{as_pretty_tibble}
\title{Create a pretty formatted tibble from a gtsummary object}
\usage{
as_pretty_tibble(summary_obj)
}
\arguments{
\item{summary_obj}{gtsummary object}
}
\value{
A tibble
}
\description{
Create a pretty formatted tibble from a gtsummary object
}
\examples{
\dontrun{
pretty_tibble(mysummary_table)
}
}
