% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_exact_ci.R
\name{binom_exact_ci}
\alias{binom_exact_ci}
\title{Calculate string confidence interval for binom exact CI}
\usage{
binom_exact_ci(num, denom, accuracy = 0.1, ci_only = FALSE, per = 1)
}
\arguments{
\item{num}{Vector number of events.}

\item{denom}{Vector size of population.}

\item{accuracy}{Accuracy of rounding}

\item{ci_only}{Boolean of whether to report CI only without mean}

\item{per}{counts per x}
}
\value{
A numerical vector.
}
\description{
Calculate string confidence interval for binom exact CI
}
\examples{
binom_exact_ci(15,80, accuracy=0.01)
}
