% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_exact_ci.R
\name{binom_exact_lower}
\alias{binom_exact_lower}
\title{Calculate lower bound of exact CI for binomial proportion}
\usage{
binom_exact_lower(num, denom)
}
\arguments{
\item{num}{Vector number of events.}

\item{denom}{Vector size of population.}
}
\value{
A numerical vector.
}
\description{
Calculate lower bound of exact CI for binomial proportion
}
\examples{
binom_exact_lower(15,80)
}
