% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fe_pooled_rr.R
\name{calculate_fe_pooled_rr}
\alias{calculate_fe_pooled_rr}
\title{Calculate fixed effects pooled risk ratio (following Borenstein 2009)}
\usage{
calculate_fe_pooled_rr(
  s1x1,
  s1n1,
  s1x2,
  s1n2,
  s2x1,
  s2n1,
  s2x2,
  s2n2,
  pretty = TRUE
)
}
\arguments{
\item{s1x1}{number of exposed cases in study 1}

\item{s1n1}{number exposed in study 1}

\item{s1x2}{number of unexposed cases in study 1}

\item{s1n2}{number unexposed in study 1}

\item{s2x1}{number of exposed cases in study 2}

\item{s2n1}{number exposed in study 2}

\item{s2x2}{number of unexposed cases in study 2}

\item{s2n2}{number unexposed in study 2}

\item{pretty}{logical indicator of whether to output a formatted string or numerical vector (default=TRUE)}
}
\value{
A numerical vector with estimate, lower bound, and upper bound or a character representation of the vector
}
\description{
Calculate fixed effects pooled risk ratio (following Borenstein 2009)
}
\examples{
calculate_fe_pooled_rr(38,398,37,514,29,314,18,215)
}
