% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx_helpers.R
\name{convert_numeric_to_binary}
\alias{convert_numeric_to_binary}
\title{Convert vector to logical if a binary number (useful for preparing for tbl_summary)}
\usage{
convert_numeric_to_binary(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Convert vector to logical if a binary number (useful for preparing for tbl_summary)
}
\examples{
\dontrun{
convert_numeric_to_binary(c(1,2,3))
}
}
