% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formatted_date.R
\name{get_formatted_date}
\alias{get_formatted_date}
\title{Get formatted current date}
\usage{
get_formatted_date()
}
\value{
A scalar string with currrent date (e.g., 20241231)
}
\description{
Get formatted current date
}
\examples{
get_formatted_date()
}
