% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx_helpers.R
\name{insertDescriptiveImage}
\alias{insertDescriptiveImage}
\title{Insert Image with a description}
\usage{
insertDescriptiveImage(..., description)
}
\arguments{
\item{...}{Arguments to insertImage function}

\item{description}{A string with the image description}
}
\value{
A numerical vector.
}
\description{
Insert Image with a description
}
\examples{
\dontrun{
InsertDescriptiveImage(wb, 1, "output/myimage.jpeg")
}
}
