% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbutils-package.R
\docType{package}
\name{jbutils-package}
\alias{jbutils}
\alias{jbutils-package}
\title{jbutils: Utility package for Jeremy Brown}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
