% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk.R
\name{riskscoreci}
\alias{riskscoreci}
\title{Calculate risk ratio and score confidence intervals (based on https://users.stat.ufl.edu/~aa/cda/R/two-sample/R2/ and epitools package)}
\usage{
riskscoreci(x1, n1, x2, n2, conflev)
}
\arguments{
\item{x1}{Scalar number exposed with outcome}

\item{n1}{Scalar number exposed}

\item{x2}{Scalar number unexposed with outcome}

\item{n2}{Scalar number unexposed}

\item{conflev}{Confidence level}
}
\value{
A vector with the lb, estimate, and ub
}
\description{
Calculate risk ratio and score confidence intervals (based on https://users.stat.ufl.edu/~aa/cda/R/two-sample/R2/ and epitools package)
}
\examples{
binom_exact_ci(15,80, accuracy=0.01)
}
