% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_score_ci.R
\name{rr_score_ci}
\alias{rr_score_ci}
\title{Calculate string confidence interval for binom exact CI}
\usage{
rr_score_ci(x1, n1, x2, n2, accuracy = 0.01, ci_only = FALSE)
}
\arguments{
\item{accuracy}{Accuracy of rounding}

\item{ci_only}{Boolean of whether to report CI only without mean}

\item{num}{Vector number of events.}

\item{denom}{Vector size of population.}
}
\value{
A numerical vector.
}
\description{
Calculate string confidence interval for binom exact CI
}
\examples{
rr_score_ci(3,50, 6, 100)
}
