% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk.R
\name{rr_score_lower}
\alias{rr_score_lower}
\title{Calculate lower bound of risk ratio score CI}
\usage{
rr_score_lower(x1, n1, x2, n2)
}
\arguments{
\item{x1}{Scalar number exposed with outcome}

\item{n1}{Scalar number exposed}

\item{x2}{Scalar number unexposed with outcome}

\item{n2}{Scalar number unexposed}
}
\value{
A numerical vector.
}
\description{
Calculate lower bound of risk ratio score CI
}
\examples{
rr_score_lower(15,80,30,160)
}
