% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk.R
\name{rr_score_upper}
\alias{rr_score_upper}
\title{Calculate upper bound of risk ratio score CI}
\usage{
rr_score_upper(x1, n1, x2, n2)
}
\arguments{
\item{x1}{Scalar number exposed with outcome}

\item{n1}{Scalar number exposed}

\item{x2}{Scalar number unexposed with outcome}

\item{n2}{Scalar number unexposed}
}
\value{
A numerical vector.
}
\description{
Calculate upper bound of risk ratio score CI
}
\examples{
rr_score_upper(15,80,30,160)
}
