% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx_helpers.R
\name{set_wb_options}
\alias{set_wb_options}
\title{Set options for workbook}
\usage{
set_wb_options(wb)
}
\arguments{
\item{wb}{Workbook object}
}
\description{
Set options for workbook
}
\examples{
\dontrun{
set_wb_options(wb)
}
}
