% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_tibble.R
\name{transpose_tibble}
\alias{transpose_tibble}
\title{Transpose a tibble}
\usage{
transpose_tibble(tibble_data, new_colname = "name")
}
\arguments{
\item{tibble_data}{Tibble or data frame with data}

\item{new_colname}{String new column name}
}
\value{
A tibble.
}
\description{
Transpose a tibble
}
\examples{
animals <- data.frame(animal=c("horse", "cheetah"), max_speed=c(44, 75), max_height=c(7, 4))
transpose_tibble(animals, "Characteristic")
}
