% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fe_pooled_rr.R
\name{var_log_rr}
\alias{var_log_rr}
\title{Calculate variance of log risk ratio (following Jewell 2003)}
\usage{
var_log_rr(x1, n1, x2, n2)
}
\description{
Calculate variance of log risk ratio (following Jewell 2003)
}
