% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson_ci.R
\name{wilson_ci}
\alias{wilson_ci}
\title{Calculate string confidence interval for wilson CI}
\usage{
wilson_ci(num, denom, accuracy = 0.1, ci_only = FALSE, per = 1)
}
\arguments{
\item{num}{Scalar number of events.}

\item{denom}{Scalar size of population.}

\item{accuracy}{Accuracy of rounding}

\item{ci_only}{Boolean of whether to report CI only without mean}

\item{per}{counts per x}
}
\value{
A numerical vector.
}
\description{
Calculate string confidence interval for wilson CI
}
\examples{
wilson_ci(15,80, accuracy=0.01)
}
