% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson_ci.R
\name{wilson_lower}
\alias{wilson_lower}
\title{Calculate lower bound of Wilson CI for binomial proportion}
\usage{
wilson_lower(num, denom)
}
\arguments{
\item{num}{Scalar number of events.}

\item{denom}{}
}
\value{
A numerical vector.
}
\description{
Calculate lower bound of Wilson CI for binomial proportion
}
\examples{
wilson_lower(15,80)
}
