% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson_ci.R
\name{wilson_upper}
\alias{wilson_upper}
\title{Calculate upper bound of Wilson CI for binomial proportion}
\usage{
wilson_upper(num, denom)
}
\arguments{
\item{num}{Scalar number of events.}

\item{denom}{Scalar size of population.}
}
\value{
A numerical vector.
}
\description{
Calculate upper bound of Wilson CI for binomial proportion
}
\examples{
wilson_upper(15,80)
}
