% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx_helpers.R
\name{writeDataTableAuto}
\alias{writeDataTableAuto}
\title{Write formatted DataTable}
\usage{
writeDataTableAuto(...)
}
\arguments{
\item{...}{Arguments to WriteDataTable function}
}
\description{
Write formatted DataTable
}
\examples{
\dontrun{
writeDataTableAuto(wb, 1, df)
}
}
