% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx_helpers.R
\name{writeDescriptiveDataTableAuto}
\alias{writeDescriptiveDataTableAuto}
\title{Write formattedDataTable with a description}
\usage{
writeDescriptiveDataTableAuto(..., description)
}
\arguments{
\item{...}{Arguments to WriteDataTable function}

\item{description}{A string with the table description}
}
\value{
A numerical vector.
}
\description{
Write formattedDataTable with a description
}
\examples{
\dontrun{
writeDescriptiveDataTableAuto(wb, 1, df)
}
}
