% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk.R
\name{rr_score_ci}
\alias{rr_score_ci}
\title{Calculate string confidence interval for risk ratio score CI}
\usage{
rr_score_ci(x1, n1, x2, n2, accuracy = 0.01, ci_only = FALSE)
}
\arguments{
\item{x1}{Scalar number exposed with outcome}

\item{n1}{Scalar number exposed}

\item{x2}{Scalar number unexposed with outcome}

\item{n2}{Scalar number unexposed}

\item{accuracy}{Accuracy of rounding}

\item{ci_only}{Boolean of whether to report CI only without mean}
}
\value{
A numerical vector.
}
\description{
Calculate string confidence interval for risk ratio score CI
}
\examples{
rr_score_ci(3,50, 6, 100)
}
