% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_score_ci.R
\name{rr_score_lower}
\alias{rr_score_lower}
\title{Calculate lower bound of risk ratio score CI}
\usage{
rr_score_lower(x1, n1, x2, n2)
}
\arguments{
\item{num}{Vector number of events.}

\item{denom}{Vector size of population.}
}
\value{
A numerical vector.
}
\description{
Calculate lower bound of risk ratio score CI
}
\examples{
rr_score_lower(15,80)
}
