% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk.R
\name{var_log_rr}
\alias{var_log_rr}
\title{Calculate variance of log risk ratio (following Jewell 2003)}
\usage{
var_log_rr(x1, n1, x2, n2)
}
\arguments{
\item{x1}{number of cases in group 1}

\item{n1}{number in group 1}

\item{x2}{number of cases in group 2}

\item{n2}{number in group 2}
}
\value{
A numerical output of estimated variance
}
\description{
Calculate variance of log risk ratio (following Jewell 2003)
}
\examples{
calculate_fe_pooled_rr(38,398,37,514,29,314,18,215)
}
